backend-chat/requirements.txt
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.5
docutils==0.21.2
fastapi==0.116.1
filelock==3.18.0
h2==4.2.0
HTMLParser==0.0.2
ipython==9.4.0
ipywidgets==8.1.7
jnius==1.1.0
js==1.0
keyring==25.6.0
openai==1.97.0
Pillow==11.3.0
protobuf==6.31.1
pydantic==2.11.7
pyodide==0.0.2
pyOpenSSL==25.1.0
pytest==8.4.1
python-dotenv==1.1.1
redis==6.2.0
simplejson==3.20.1
Sphinx==8.2.3
thread==2.0.5
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zstandard==0.23.0
